+++ example/lib/inputs/go_dynamic_input.dart
@@ -1,10 +1,16 @@
-
 import 'package:flutter/material.dart';
 import 'package:go_form/go_form.dart';
 
-class GoDynamicInput extends FormFieldModelBase<String>{
+class GoDynamicInput extends FormFieldModelBase<String> {
   final String label;
-  GoDynamicInput( {required super.name, super.validator,required this.label,super.initialValue });
+
+  GoDynamicInput({
+    required super.name,
+    super.validator,
+    required this.label,
+    super.initialValue,
+    super.key,
+  });
 
   @override
   Widget build(BuildContext context, FieldController<String> controller) {